organize files in module structure
