+++ designatedlands.cfg
@@ -1,0 +1,10 @@
+# Change default database/paths/filenames etc here (see README.md)
+[designatedlands]
+source_data=source_data
+source_csv=sources.csv
+out_table=designatedlands
+out_file=designatedlands.gpkg
+out_format=GPKG
+db_url=postgresql://postgres:postgres@localhost:5432/designatedlands
+n_processes=-1
+email=
