+++ designatedlands/__init__.py
@@ -1,0 +1,5 @@
+from .designatedlands import *
+from . import util
+from . import download
+
+__version__ = "0.1.0"
