+++ tests/test.cfg
@@ -1,10 +1,26 @@
-# Change default database/paths/filenames etc here (see README.md)
-
 [designatedlands]
+
+# email for DataBC downloads
+# (if blank, defaults to $BCDATA_EMAIL environment variable)
+email=
+
+# data source paths
 dl_path=tests/source_data
 source_csv=tests/sources_test.csv
-out_table=tests/designatedlands
+
+# output names/paths
+out_table=designatedlands
 out_file=tests/designatedlands.gpkg
 out_format=GPKG
+
+# processing config
 db_url=postgresql://postgres:postgres@localhost:5432/designatedlands_test
-n_processes=0
+n_processes=-1
+
+# logging config
+log_level = 10           # 10:DEBUG, 20:INFO, 30:WARNING, 40:ERROR, 50:CRITICAL
+log_file = True
+log_console = False
+log_name = designatedlands
+log_filename = designatedlands
+logs_folder = logs
